pyproject.toml
[tool.poetry]
name = "amplitude-collector"
version = "0.2.2"
description = ""
authors = ["Andrey Tatarinov <a@tatarinov.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.21.0"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
kafka-python = "^2.0.2"
pytest = "^7.1.3"
httpx = "^0.23.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.982"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
